// src/shared/components/configurable-parameters/configurable-parameters.interface.ts
export interface ConfigurableParametersComponents
    extends Omit<ConfigurableParametersComponentsDTO, 'groups' | 'parameters' | 'entity_identifier'> {
    entityIdentifier: EntityIdentifier;
    groups?: ConfigurableParametersGroups[];
    parameters?: ConfigurableParametersParams[];
}

export interface ConfigurableParametersTaskChain {
    taskId: ConfigurableParametersTaskChainDTO['task_id'];
    taskTitle: ConfigurableParametersTaskChainDTO['task_title'];
    components: ConfigurableParametersComponents[];
}

export enum ConfigurableParametersType {
    SINGLE_CONFIG_PARAMETERS = 'single-config-parameters',
    MANY_CONFIG_PARAMETERS = 'many-config-parameters',
    READ_ONLY_SINGLE_PARAMETERS = 'read-only-single-parameters',
}

export interface ConfigurableParametersSingle {
    type: ConfigurableParametersType.SINGLE_CONFIG_PARAMETERS;
    configParametersData: ConfigurableParametersTaskChain;
    updateParameter<T extends string | boolean | number>(id: string, value: T): void;
}

interface ConfigurableParametersSingleReadOnly {
    type: ConfigurableParametersType.READ_ONLY_SINGLE_PARAMETERS;
    configParametersData: ConfigurableParametersTaskChain;
}

export interface ConfigurableParametersMany {
    type: ConfigurableParametersType.MANY_CONFIG_PARAMETERS;
    configParametersData: ConfigurableParametersTaskChain[];
    updateParameter: ConfigurableParametersSingle['updateParameter'];
    selectedComponentId: string | undefined;
    setSelectedComponentId: Dispatch<SetStateAction<string | undefined>>;
    selectedComponent: ConfigurableParametersComponents | undefined;
}

export type ConfigurableParametersProps =
    | (ConfigurableParametersSingle & { hpo: HPOProps })
    | ConfigurableParametersSingleReadOnly
    | ConfigurableParametersMany;